requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<3.0
django-redis>=5.4,<5.5
celery>=5.3,<5.4
redis>=5.0,<5.1